apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.12.0
boto3==1.34.46
python-magic==0.4.27
slowapi==0.1.9
httpx==0.26.0
pytest==8.0.1
pytest-asyncio==0.23.5
minio==7.2.4
python-json-logger==2.0.7
structlog==24.1.0
aiosqlite==0.19.0
email-validator==2.1.0.post1
tenacity==8.2.3
